Profile: put dark mode outside and simplify login
