+++ src/components/profile/AccountPopover.jsx
@@ -2,6 +2,7 @@
 
 import React from "react";
 import {
+  Button,
   IconButton,
   Avatar,
   Popover,
@@ -18,21 +19,14 @@ import Icon from "components/Icon";
 import { getFile } from "utils/files";
 import { login, logout } from "utils/auth";
 import { useAuth } from "components/AuthProvider";
-import { useMode, useMode2 } from "contexts/ModeContext";
-import { ModeSwitch } from "components/ModeSwitch";
+import { useMode2 } from "contexts/ModeContext";
 import { usePathname } from "next/navigation";
 
 export default function AccountPopover() {
   const pathname = usePathname();
   const { user, isAuthenticated } = useAuth();
-  const { isDark, setMode } = useMode(); // Accessing isDark and setMode from ModeContext
   const { isiframe, setIsiframe } = useMode2();
   const [open, setOpen] = React.useState(null);
-
-  const handleChange = () => {
-    // handleupdate();
-    setMode(!isDark); // Toggle the mode
-  };
 
   const handleOpen = (event) => {
     setOpen(event.currentTarget);
@@ -50,77 +44,87 @@ export default function AccountPopover() {
     },
   ];
 
-  const COMMON_MENU_OPTIONS = [];
+  if (isiframe) return null;
+
+  if (!isAuthenticated) {
+    return (
+      <Button
+        sx={{
+          minWidth: "80px",
+          padding: "10px",
+        }}
+        variant="outlined"
+        onClick={() => login(pathname)}
+      >
+        LOGIN
+      </Button>
+    );
+  }
 
   return (
     <>
-      {!isiframe && (
-        <IconButton
-          onClick={handleOpen}
-          aria-label="Account Popover"
-          sx={{
-            p: 0,
-            color: "white",
-            ...(open && {
-              "&:before": {
-                zIndex: 1,
-                content: "''",
-                width: "100%",
-                height: "100%",
-                borderRadius: "50%",
-                position: "absolute",
-                bgcolor: (theme) =>
-                  alpha(theme.palette.background.neutral, 0.5),
-              },
-            }),
-            ...(!open && {
-              // "&:after": {
-              "&:before": {
-                zIndex: 1,
-                content: "''",
-                width: "100%",
-                height: "100%",
-                borderRadius: "50%",
-                position: "absolute",
-                bgcolor: (theme) =>
-                  alpha(theme.palette.background.neutral, 0.1),
-              },
-              // },
-            }),
-          }}
+      <IconButton
+        onClick={handleOpen}
+        aria-label="Account Popover"
+        sx={{
+          color: "white",
+          ...(open && {
+            "&:before": {
+              zIndex: 1,
+              content: "''",
+              width: "100%",
+              height: "100%",
+              borderRadius: "50%",
+              bgcolor: (theme) =>
+                alpha(theme.palette.background.neutral, 0.5),
+            },
+          }),
+          ...(!open && {
+            // "&:after": {
+            "&:before": {
+              zIndex: 1,
+              content: "''",
+              width: "100%",
+              height: "100%",
+              borderRadius: "50%",
+              bgcolor: (theme) =>
+                alpha(theme.palette.background.neutral, 0.1),
+            },
+            // },
+          }),
+        }}
+      >
+        <Avatar
+          width={40}
+          height={40}
+          {...(user?.firstName && {
+            children: `${user?.firstName?.[0]}${
+              user?.lastName == "" ? "" : user?.lastName?.[0]
+            }`,
+          })}
         >
-          <Avatar
-            width={40}
-            height={40}
-            {...(user?.firstName && {
-              children: `${user?.firstName?.[0]}${
-                user?.lastName == "" ? "" : user?.lastName?.[0]
-              }`,
-            })}
-          >
-            {user?.img ? (
-              <Image
-                alt={user?.firstName}
-                src={getFile(user?.img)}
-                width={400}
-                height={400}
-                style={{
-                  top: 0,
-                  left: 0,
-                  width: "100%",
-                  height: "100%",
-                  objectFit: "cover",
-                  position: "absolute",
-                }}
-              />
-            ) : user?.firstName ? (
-              `${user?.firstName?.[0]}${
-                user?.lastName === "" ? "" : user?.lastName?.[0]
-              }`
-            ) : null}
-          </Avatar>
-        </IconButton>
-      )}
+          {user?.img ? (
+            <Image
+              alt={user?.firstName}
+              src={getFile(user?.img)}
+              width={400}
+              height={400}
+              style={{
+                top: 0,
+                left: 0,
+                width: "100%",
+                height: "100%",
+                objectFit: "cover",
+                position: "absolute",
+              }}
+            />
+          ) : user?.firstName ? (
+            `${user?.firstName?.[0]}${
+              user?.lastName === "" ? "" : user?.lastName?.[0]
+            }`
+          ) : null}
+        </Avatar>
+      </IconButton>
 
       <Popover
         open={Boolean(open)}
@@ -141,59 +145,27 @@ export default function AccountPopover() {
           },
         }}
       >
-        <ModeSwitch checked={isDark} onChange={handleChange} sx={{ m: 1 }} />
+        <>
+          <Stack sx={{ my: 1.5, px: 2.5 }}>
+            <Typography variant="subtitle2" noWrap>
+              {`${user?.firstName} ${user?.lastName}`}
+            </Typography>
+            <Typography
+              variant="body2"
+              sx={{ color: "text.secondary" }}
+              noWrap
+            >
+              {user?.email}
+            </Typography>
+          </Stack>
 
-        {/* <ModeSwitch checked={isDark} onChange={handleChange} sx={{ m: 1 }} />{" "} */}
-        {/* Pass current isDark value and handleChange function to ModeSwitch component */}
-        {isAuthenticated ? (
-          <>
-            <Stack sx={{ my: 1.5, px: 2.5 }}>
-              <Typography variant="subtitle2" noWrap>
-                {`${user?.firstName} ${user?.lastName}`}
-              </Typography>
-              <Typography
-                variant="body2"
-                sx={{ color: "text.secondary" }}
-                noWrap
-              >
-                {user?.email}
-              </Typography>
-            </Stack>
+          {[...AUTHENTICATED_MENU_OPTIONS].length > 0 ? (
+            <>
+              <Divider sx={{ borderStyle: "dashed" }} />
 
-            {[...AUTHENTICATED_MENU_OPTIONS, ...COMMON_MENU_OPTIONS].length >
-            0 ? (
-              <>
-                <Divider sx={{ borderStyle: "dashed" }} />
-
-                <Stack sx={{ p: 1 }}>
-                  {[...AUTHENTICATED_MENU_OPTIONS, ...COMMON_MENU_OPTIONS].map(
-                    (option) => (
-                      <MenuItem
-                        component={Link}
-                        key={option.label}
-                        href={option.url}
-                      >
-                        <Icon variant={option.icon} sx={{ mr: 2 }} />
-                        {option.label}
-                      </MenuItem>
-                    ),
-                  )}
-                </Stack>
-              </>
-            ) : null}
-
-            <Divider sx={{ borderStyle: "dashed" }} />
-
-            <MenuItem onClick={() => logout(pathname)} sx={{ m: 1 }}>
-              Logout
-            </MenuItem>
-          </>
-        ) : (
-          <>
-            {COMMON_MENU_OPTIONS.length > 0 ? (
-              <>
-                <Stack sx={{ p: 1 }}>
-                  {COMMON_MENU_OPTIONS.map((option) => (
+              <Stack sx={{ p: 1 }}>
+                {[...AUTHENTICATED_MENU_OPTIONS].map(
+                  (option) => (
                     <MenuItem
                       component={Link}
                       key={option.label}
@@ -202,18 +174,18 @@ export default function AccountPopover() {
                       <Icon variant={option.icon} sx={{ mr: 2 }} />
                       {option.label}
                     </MenuItem>
-                  ))}
-                </Stack>
+                  ),
+                )}
+              </Stack>
+            </>
+          ) : null}
 
-                <Divider sx={{ borderStyle: "dashed" }} />
-              </>
-            ) : null}
+          <Divider sx={{ borderStyle: "dashed" }} />
 
-            <MenuItem onClick={() => login(pathname)} sx={{ m: 1 }}>
-              Login
-            </MenuItem>
-          </>
-        )}
+          <MenuItem onClick={() => logout(pathname)} sx={{ m: 1 }}>
+            Logout
+          </MenuItem>
+        </>
       </Popover>
     </>
   );
